requirements.txt
rich>=14.2.0
psutil>=7.1.3
json5>=0.13.0
PySide6_Essentials>=6.10.0  # Addons unused
# qt-material>=2.17  Exported theme to local
pillow>=12.0.0
tqdm>=4.67.1
pywin32>=310; sys_platform == 'win32'
darkdetect>=0.8.0
